requirements.txt
pwntools==4.7.0
rich==10.15.1